[Rename] 분석 페이지 폴더 분리
