+++ src/app/Router.tsx
@@ -15,8 +15,8 @@ import VotePage from '@/pages/vote/VotePage'
 import AddVotePage from '@/pages/vote/AddVotePage'
 import EditVotePage from '@/pages/vote/EditVotePage'
 import AddItem from '@/pages/item/add/AddItem'
-import StatisticsPage from '@/pages/statistics/StatisticsPage'
-import StatisticsDetailPage from '@/pages/statistics/StatisticsDetailPage'
+import StatisticsPage from '@/pages/statistics/ui/StatisticsPage'
+import StatisticsDetailPage from '@/pages/statistics/ui/StatisticsDetailPage'
 import NotFound from '@/shared/components/notFound/NotFound'
 import FAQPage from '@/pages/faq/page'
 import NoticePage from '@/pages/notice/page'
