+++ src/features/statistics/model/constant.ts
