+++ src/pages/statistics/ui/index.ts
@@ -1,0 +1,2 @@
+export { StatisticsDetailPage } from './StatisticsDetailPage'
+export { StatisticsPage } from './StatisticsPage'
